+++ packages/react-ui/components/FxInput/FxInput.tsx
@@ -1,5 +1,6 @@
 import type { AriaAttributes } from 'react';
 import React from 'react';
+import warning from 'warning';
 
 import { Group } from '../Group';
 import type { InputProps } from '../Input';
@@ -16,6 +17,11 @@ import { rootNode } from '../../lib/rootNode';
 import { ThemeContext } from '../../lib/theming/ThemeContext';
 import type { Theme } from '../../lib/theming/Theme';
 import type { SizeProp } from '../../lib/types/props';
+import type { MaskedInputProps } from '../MaskedInput';
+import { MaskedInput } from '../MaskedInput';
+import type { ReactUIFeatureFlags } from '../../lib/featureFlagsContext';
+import { getFullReactUIFlagsContext } from '../../lib/featureFlagsContext';
+import { ReactUIFeatureFlagsContext } from '../../lib/featureFlagsContext/ReactUIFeatureFlagsContext';
 import { withSize } from '../../lib/size/SizeDecorator';
 
 import { MathFunctionIcon } from './MathFunctionIcon';
@@ -24,8 +30,9 @@ import { FxInputRestoreBtn } from './FxInputRestoreBtn';
 export interface FxInputProps
   extends Pick<AriaAttributes, 'aria-label'>,
     CommonProps,
+    Pick<Partial<MaskedInputProps>, 'alwaysShowMask' | 'mask' | 'maskChar' | 'formatChars'>,
     Override<
-      CurrencyInputProps,
+      Omit<CurrencyInputProps, 'alwaysShowMask' | 'mask' | 'maskChar' | 'formatChars'>,
       {
         /** Устанавливает авто-режим. */
         auto?: boolean;
@@ -60,7 +67,7 @@ export const FxInputDataTids = {
   root: 'FxInput__root',
 } as const;
 
-type DefaultProps = Required<Pick<FxInputProps, 'width' | 'type' | 'value'>>;
+type DefaultProps = Required<Pick<FxInputProps, 'width' | 'type' | 'value' | 'auto'>>;
 type DefaultizedFxInputProps = DefaultizedProps<FxInputProps, DefaultProps>;
 
 /**
@@ -81,6 +88,7 @@ export class FxInput extends React.Component<FxInputProps> {
     width: 250,
     type: 'text',
     value: '',
+    auto: false,
   };
 
   private theme!: Theme;
@@ -90,82 +98,61 @@ export class FxInput extends React.Component<FxInputProps> {
   private getProps = createPropsGetter(FxInput.defaultProps);
   public getRootNode!: TGetRootNode;
   private setRootNode!: TSetRootNode;
+  private featureFlags!: ReactUIFeatureFlags;
+
+  private validateProps(props: FxInputProps) {
+    warning(
+      props.type !== 'currency' && props.mask !== undefined,
+      '[FxInput]: Prop "mask" is not supported when type="currency"',
+    );
+  }
+
+  public componentDidMount() {
+    this.validateProps(this.getProps());
+  }
+
+  public componentDidUpdate() {
+    this.validateProps(this.getProps());
+  }
 
   public render() {
     return (
-      <ThemeContext.Consumer>
-        {(theme) => {
-          this.theme = theme;
+      <ReactUIFeatureFlagsContext.Consumer>
+        {(flags) => {
+          this.featureFlags = getFullReactUIFlagsContext(flags);
           return (
-            <CommonWrapper rootNodeRef={this.setRootNode} {...this.getProps()}>
-              {this.renderMain}
-            </CommonWrapper>
+            <ThemeContext.Consumer>
+              {(theme) => {
+                this.theme = theme;
+                return (
+                  <CommonWrapper rootNodeRef={this.setRootNode} {...this.getProps()}>
+                    {this.renderMain}
+                  </CommonWrapper>
+                );
+              }}
+            </ThemeContext.Consumer>
           );
         }}
-      </ThemeContext.Consumer>
+      </ReactUIFeatureFlagsContext.Consumer>
     );
   }
 
   public renderMain = (props: CommonWrapperRestProps<DefaultizedFxInputProps>) => {
-    const { type, onRestore, auto, refInput, value, width, ...rest } = props;
-    const inputProps: Partial<CurrencyInputProps> = {
-      align: 'right',
-    };
-
-    let button = null;
-    const inputCorners: InputProps['corners'] = auto
-      ? { ...rest.corners }
-      : { borderBottomLeftRadius: 0, borderTopLeftRadius: 0, ...rest.corners };
-    const iconSizes: Record<SizeProp, number> = {
-      small: parseInt(this.theme.inputIconSizeSmall),
-      medium: parseInt(this.theme.inputIconSizeMedium),
-      large: parseInt(this.theme.inputIconSizeLarge),
-    };
-    const size = this.size;
-    const IconFunction = <MathFunctionIcon size={iconSizes[size]} />;
-
-    if (auto) {
-      inputProps.leftIcon = IconFunction;
-    } else {
-      button = (
-        <FxInputRestoreBtn
-          size={this.size}
-          onRestore={onRestore}
-          corners={rest.corners}
-          disabled={rest.disabled}
-          borderless={rest.borderless}
-          aria-label={props.buttonAriaLabel}
-        />
-      );
-    }
+    const button = props.auto ? null : (
+      <FxInputRestoreBtn
+        size={props.size}
+        onRestore={props.onRestore}
+        corners={props.corners}
+        disabled={props.disabled}
+        borderless={props.borderless}
+        aria-label={props.buttonAriaLabel}
+      />
+    );
 
     return (
-      <Group data-tid={FxInputDataTids.root} width={width}>
+      <Group data-tid={FxInputDataTids.root} width={props.width}>
         {button}
-        {type === 'currency' ? (
-          <CurrencyInput
-            {...inputProps}
-            {...rest}
-            corners={inputCorners}
-            size={this.size}
-            width={'100%'}
-            ref={this.refInput}
-            value={value as CurrencyInputProps['value']}
-            onValueChange={this.props.onValueChange as CurrencyInputProps['onValueChange']}
-          />
-        ) : (
-          <Input
-            {...inputProps}
-            {...rest}
-            corners={inputCorners}
-            size={this.size}
-            width={'100%'}
-            ref={this.refInput}
-            type={type}
-            value={value as InputProps['value']}
-            onValueChange={this.props.onValueChange as InputProps['onValueChange']}
-          />
-        )}
+        {this.getInput(props)}
       </Group>
     );
   };
@@ -194,5 +181,96 @@ export class FxInput extends React.Component<FxInputProps> {
     if (this.props.refInput) {
       this.props.refInput(this.input);
     }
+  };
+
+  private getLeftIcon = (props: FxInputProps) => {
+    if (!props.auto) {
+      return null;
+    }
+
+    const iconSizes: Record<SizeProp, number> = {
+      small: parseInt(this.theme.inputIconSizeSmall),
+      medium: parseInt(this.theme.inputIconSizeMedium),
+      large: parseInt(this.theme.inputIconSizeLarge),
+    };
+    const size = this.size;
+
+    return <MathFunctionIcon size={iconSizes[size]} />;
+  };
+
+  private getInput = (props: FxInputProps) => {
+    const {
+      type,
+      onRestore,
+      auto,
+      refInput,
+      value,
+      width,
+      size: _size,
+      mask,
+      maskChar,
+      formatChars,
+      alwaysShowMask,
+      signed,
+      integerDigits,
+      fractionDigits,
+      hideTrailingZeros,
+      buttonAriaLabel,
+      corners: originalCorners,
+      ...rest
+    } = props;
+    const corners: InputProps['corners'] = auto
+      ? originalCorners
+      : { borderBottomLeftRadius: 0, borderTopLeftRadius: 0, ...originalCorners };
+    const size = this.size;
+    const commonInputProps = {
+      corners,
+      size,
+      width: '100%',
+      ref: this.refInput,
+      ...rest,
+    };
+
+    if (type === 'currency') {
+      return (
+        <CurrencyInput
+          {...commonInputProps}
+          signed={signed}
+          integerDigits={integerDigits}
+          fractionDigits={fractionDigits}
+          value={value as CurrencyInputProps['value']}
+          onValueChange={this.props.onValueChange as CurrencyInputProps['onValueChange']}
+        />
+      );
+    }
+    if (this.featureFlags.fxInputUseMaskedInput && mask) {
+      return (
+        <MaskedInput
+          {...commonInputProps}
+          leftIcon={this.getLeftIcon(props)}
+          align="right"
+          mask={mask}
+          maskChar={maskChar}
+          formatChars={formatChars}
+          alwaysShowMask={alwaysShowMask}
+          value={value as MaskedInputProps['value']}
+          onValueChange={this.props.onValueChange as MaskedInputProps['onValueChange']}
+        />
+      );
+    }
+    return (
+      <Input
+        {...commonInputProps}
+        leftIcon={this.getLeftIcon(props)}
+        align="right"
+        mask={mask}
+        maskChar={maskChar}
+        formatChars={formatChars}
+        alwaysShowMask={alwaysShowMask}
+        type={type}
+        value={value as InputProps['value']}
+        onValueChange={this.props.onValueChange as InputProps['onValueChange']}
+      />
+    );
   };
 }