+++ packages/react-ui-smoke-test/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.1.1](https://git.skbkontur.ru/ui/react-ui/compare/react-ui-smoke-test@3.1.0...react-ui-smoke-test@3.1.1) (2025-11-01)
+
+**Note:** Version bump only for package react-ui-smoke-test
+
+
+
+
+
 # [3.1.0](https://git.skbkontur.ru/ui/react-ui/compare/react-ui-smoke-test@3.0.6...react-ui-smoke-test@3.1.0) (2025-10-23)