+++ packages/react-ui-validations/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.3.5](https://github.com/skbkontur/retail-ui/compare/react-ui-validations@2.3.4...react-ui-validations@2.3.5) (2025-11-01)
+
+
+### Bug Fixes
+
+* **FxInput:** add support `mask` & feature flag `fxInputUseMaskedInput` ([f86d3bd](https://github.com/skbkontur/retail-ui/commit/f86d3bde8e205dbea77faee0fd079266af04980f))
+
+
+
+
+
 ## [2.3.4](https://github.com/skbkontur/retail-ui/compare/react-ui-validations@2.3.3...react-ui-validations@2.3.4) (2025-10-23)
 
 **Note:** Version bump only for package react-ui-validations